Extract update banner to own component
